plugins/git: Further improvements
The gwc function can now also create a worktree if it does not yet
exist. When creating a new worktree (either with gwa or gwc), the
worktree can either track an existing branch (if one with this name
already exists) or the new worktree is created along with a branch with
the same name.

Signed-off-by: Bernd Finger <bfinger@redhat.com>